+++ src/main.cpp
@@ -1,5 +1,11 @@
 #include <Geode/Geode.hpp>
 using namespace geode::prelude;
 
+
+/*
+ *   If you're a developer and wish to get a list of this mod's 
+ *   node IDs, check core/node_ids.h for the complete list.
+ */
+
 #include "hooks/level_info_layer_hook.h"
 #include "hooks/edit_level_layer_hook.h"